# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flujo de Caja",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/imports/plantilla.csv": {
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "Download Template",
        "operationId": "download_template_api_v1_imports_plantilla_csv_get",
        "parameters": [
          {
            "name": "tipo_importacion",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "REAL",
                "PROYECCION"
              ],
              "type": "string",
              "default": "REAL",
              "title": "Tipo Importacion"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/csv": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Import Csv",
        "operationId": "import_csv_api_v1_imports_csv_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_csv_api_v1_imports_csv_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/imports/{batch_id}": {
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "Get Batch",
        "operationId": "get_batch_api_v1_imports__batch_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportBatchRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_csv_api_v1_imports_csv_post": {
        "properties": {
          "cuenta_bancaria_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cuenta Bancaria Id"
          },
          "archivo": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Archivo"
          },
          "tipo_importacion": {
            "type": "string",
            "enum": [
              "REAL",
              "PROYECCION"
            ],
            "title": "Tipo Importacion",
            "default": "REAL"
          }
        },
        "type": "object",
        "required": [
          "cuenta_bancaria_id",
          "archivo"
        ],
        "title": "Body_import_csv_api_v1_imports_csv_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportBatchRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "cuenta_bancaria_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cuenta Bancaria Id"
          },
          "archivo_nombre": {
            "type": "string",
            "title": "Archivo Nombre"
          },
          "usuario_id": {
            "type": "string",
            "format": "uuid",
            "title": "Usuario Id"
          },
          "fecha_importacion": {
            "type": "string",
            "format": "date-time",
            "title": "Fecha Importacion"
          },
          "total_filas": {
            "type": "integer",
            "title": "Total Filas"
          },
          "filas_nuevas": {
            "type": "integer",
            "title": "Filas Nuevas"
          },
          "filas_duplicadas": {
            "type": "integer",
            "title": "Filas Duplicadas"
          },
          "filas_error": {
            "type": "integer",
            "title": "Filas Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "cuenta_bancaria_id",
          "archivo_nombre",
          "usuario_id",
          "fecha_importacion",
          "total_filas",
          "filas_nuevas",
          "filas_duplicadas",
          "filas_error"
        ],
        "title": "ImportBatchRead"
      },
      "ImportDuplicateDetail": {
        "properties": {
          "fila": {
            "type": "integer",
            "title": "Fila"
          },
          "n_operacion": {
            "type": "string",
            "title": "N Operacion"
          },
          "motivo": {
            "type": "string",
            "title": "Motivo"
          }
        },
        "type": "object",
        "required": [
          "fila",
          "n_operacion",
          "motivo"
        ],
        "title": "ImportDuplicateDetail"
      },
      "ImportErrorDetail": {
        "properties": {
          "fila": {
            "type": "integer",
            "title": "Fila"
          },
          "mensaje": {
            "type": "string",
            "title": "Mensaje"
          }
        },
        "type": "object",
        "required": [
          "fila",
          "mensaje"
        ],
        "title": "ImportErrorDetail"
      },
      "ImportResponse": {
        "properties": {
          "batch_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          },
          "total_filas": {
            "type": "integer",
            "title": "Total Filas"
          },
          "filas_nuevas": {
            "type": "integer",
            "title": "Filas Nuevas"
          },
          "filas_duplicadas": {
            "type": "integer",
            "title": "Filas Duplicadas"
          },
          "filas_error": {
            "type": "integer",
            "title": "Filas Error"
          },
          "duplicados": {
            "items": {
              "$ref": "#/components/schemas/ImportDuplicateDetail"
            },
            "type": "array",
            "title": "Duplicados"
          },
          "errores": {
            "items": {
              "$ref": "#/components/schemas/ImportErrorDetail"
            },
            "type": "array",
            "title": "Errores"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "total_filas",
          "filas_nuevas",
          "filas_duplicadas",
          "filas_error",
          "duplicados",
          "errores"
        ],
        "title": "ImportResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  }
}